refactor: add generateStaticFunctions to reduce boilerplate
