+++ src/pages/vault/tutorials/[collectionSlug].tsx
@@ -1,45 +1,12 @@
-import { GetStaticPropsResult } from 'next'
-import vaultData from 'data/vault.json'
 import { LearnProductData } from 'types/products'
-import { ProductOption } from 'lib/learn-client/types'
 import CollectionView from 'views/collection-view'
-import {
-	getCollectionPageProps,
-	getCollectionPaths,
-	CollectionPageProps,
-} from 'views/collection-view/server'
+import { generateStaticFunctions } from 'views/collection-view/server'
+// product data
+import vaultData from 'data/vault.json'
 
-export async function getStaticProps({
-	params,
-}): Promise<GetStaticPropsResult<CollectionPageProps>> {
-	const { collectionSlug } = params
-	const product = vaultData as LearnProductData
-
-	const props = await getCollectionPageProps(product, collectionSlug)
-
-	// If the collection doesn't exist, hit the 404
-	if (!props) {
-		return { notFound: true }
-	}
-
-	return props
-}
-
-interface CollectionPagePaths {
-	params: {
-		collectionSlug: string
-	}
-}
-
-export async function getStaticPaths(): Promise<{
-	paths: CollectionPagePaths[]
-	fallback: boolean
-}> {
-	const paths = await getCollectionPaths(ProductOption['vault'])
-	return {
-		paths,
-		fallback: false,
-	}
-}
+const { getStaticPaths, getStaticProps } = generateStaticFunctions(
+	vaultData as LearnProductData
+)
+export { getStaticPaths, getStaticProps }
 
 export default CollectionView
